v1.0.2
The code was improved by adding 2 dictionaries, one for the paths and another for the clues, the code is now cleaner and tidier. 
The attribute of tiredness was added and the conditionals of some functions were also improved to show the events in Romualdo’s life in a better way.